Moved WindowManager class
