+++ src/guis/krow/pages/ScrollMenu.java
@@ -9,7 +9,7 @@ import javafx.scene.input.ScrollEvent.VerticalTextScrollUnits;
 import javafx.scene.layout.Pane;
 import krow.scene.HorizontalScrollBox;
 import kröw.core.Kröw;
-import kröw.core.managers.WindowManager.App;
+import kröw.gui.WindowManager.App;
 
 public abstract class ScrollMenu extends App {
