+++ src/guis/zeale/guis/Home.java
@@ -19,9 +19,9 @@ import krow.guis.PopupHelper;
 import krow.guis.ShapeFactory;
 import krow.pages.ScrollMenu;
 import kröw.core.Kröw;
-import kröw.core.managers.WindowManager;
-import kröw.core.managers.WindowManager.NotSwitchableException;
-import kröw.core.managers.WindowManager.App;
+import kröw.gui.WindowManager;
+import kröw.gui.WindowManager.App;
+import kröw.gui.WindowManager.NotSwitchableException;
 import zeale.guis.developer_page.DeveloperApp;
 import zeale.guis.math_app.MathApp;
 import zeale.guis.math_app.controllers.Calculator;
