+++ src/guis/zeale/guis/Tools.java
@@ -41,8 +41,8 @@ import javafx.scene.text.FontWeight;
 import krow.guis.GUIHelper;
 import krow.guis.PopupHelper;
 import kröw.core.Kröw;
-import kröw.core.managers.WindowManager;
-import kröw.core.managers.WindowManager.App;
+import kröw.gui.WindowManager;
+import kröw.gui.WindowManager.App;
 import sun.awt.shell.ShellFolder;
 
 public class Tools extends App {
